[RFR] move journal_entry_ref to account_payment_include_draft_move
