+++ account_banking_payment_export/models/account_move_line.py
@@ -25,20 +25,6 @@ from openerp import models, fields, api
 class AccountMoveLine(models.Model):
     _inherit = 'account.move.line'
 
-    @api.one
-    @api.depends('move_id.state', 'move_id.name', 'invoice.number')
-    def _get_journal_entry_ref(self):
-        if self.move_id.state == 'draft':
-            if self.invoice.id:
-                self.journal_entry_ref = self.invoice.number
-            else:
-                self.journal_entry_ref = '*' + str(self.move_id.id)
-        else:
-            self.journal_entry_ref = self.move_id.name
-
-    journal_entry_ref = fields.Char(compute=_get_journal_entry_ref,
-                                    string='Journal Entry Ref', store=True)
-
     @api.multi
     def get_balance(self):
         """
