+++ account_payment_include_draft_move/__init__.py
@@ -21,3 +21,4 @@
 ##############################################################################
 
 from . import wizard
+from . import models
